% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFrequencyDf.R
\name{buildFrequencyDf}
\alias{buildFrequencyDf}
\title{Calculates coefficients for real-valued fourier transform of an image.}
\usage{
buildFrequencyDf(dfY)
}
\arguments{
\item{dfY}{a data.frame representation of an image.}
}
\value{
a data.frame with rows corresponding to combinations of frequencies and functions and their respective real-valued fourier transform coefficient.
}
\description{
Calculates coefficients for real-valued fourier transform of an image.
}
\author{
Victor Freguglia Souza
}
