% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorMap.R
\name{CorMap}
\alias{CorMap}
\title{Correlation Map}
\usage{
CorMap(IMG, window_size = 10)
}
\arguments{
\item{IMG}{An image in a matrix object.}

\item{window_size}{Size of the windows of relative positions com compute correlation.}
}
\value{
A data.frame object with relative positions and absolute auto-correlation of each relative position
}
\description{
Computes Correlation Map
}
\author{
Victor Freguglia Souza
}
