% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DesignMat.R
\name{DesignMat}
\alias{DesignMat}
\title{Creates design Matrix for a selected set of fourier basis.}
\usage{
DesignMat(coeftab, dime)
}
\arguments{
\item{coeftab}{A filtered list of fourier basis. Result from betaPLS function.}

\item{dim}{Dimensions of the image.}
}
\value{
a design matrix
}
\description{
Creates design Matrix for a selected set of fourier basis.
}
\author{
Victor Freguglia Souza
}
