% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifHistogram.R
\name{DifHistogram}
\alias{DifHistogram}
\title{Gray-level difference histogram}
\usage{
DifHistogram(X, positions)
}
\arguments{
\item{X}{the observed field (matrix object).}

\item{positions}{either the (C x 2) matrix of the relative positions considered or a GibbsModel object.}
}
\description{
Creates the histogram of gray-level difference for each relative position in a given field.
}
\author{
Victor Freguglia Souza
}
