% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cMat.R
\name{extract_cMat}
\alias{extract_cMat}
\title{Extract the cMat structure for a Thresholded Interaction Map}
\usage{
extract_cMat(iMap)
}
\arguments{
\item{iMap}{a (thresholded) Interaction Map.}
}
\value{
a Matrix with the interacting pairs.
}
\description{
Extract the cMat structure for a Thresholded Interaction Map
}
\author{
Victor Freguglia Souza
}
