% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsRF_utilities.R
\name{getFrequencyCoef2}
\alias{getFrequencyCoef2}
\title{Gets real-valued fourier transform coefficent for a specific pair of frequencies and functions based on the (matrix) fft.}
\usage{
getFrequencyCoef2(ft, n, m, f1, f2)
}
\arguments{
\item{ft}{the fourier transform (matrix) object of a image.}

\item{n, m}{the frequency to get coefficients for.}

\item{f1, f2}{names of the functions for the coefficient, can be either "cos" or "sin".}
}
\value{
the numeric value of the coefficient.
}
\description{
Gets real-valued fourier transform coefficent for a specific pair of frequencies and functions based on the (matrix) fft.
}
\author{
Victor Freguglia Souza
}
