% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsMCMLE.R
\name{GibbsMCMLE}
\alias{GibbsMCMLE}
\title{Monte-Carlo Maximum Likelihood Estimate of Potentials}
\usage{
GibbsMCMLE(X, gModelX, type = "symetric", gModelRef = "MPLE",
  nsamples = 100, showProgress = TRUE, ...)
}
\arguments{
\item{X}{the random field (matrix object).}

\item{gModelX}{a GibbsModel for X (V and vMat can be NULL).}

\item{type}{Type of function for Potentials. See GibbsMPLE() for reference.}

\item{gModelRef}{the reference model to simulate from for Monte-Carlo Likelihood function approximation. Either a GibbsModel object or a character with "zero" for an Independent Random Field, or "MPLE" for MPL estimated model.}

\item{nsamples}{an integer indicating the number of Monte-Carlo samples.}

\item{showProgress}{indicates whether the current step being done should be printed or not.}

\item{...}{additional parameters to be used in the optimization function (optim()).}
}
\value{
a GibbsModel object with the estimated potentials.
}
\description{
Provides MC Maximum Likelihood Estimation of Potentials for a given image.
}
\examples{
GibbsMCMLE(example.X,example.GibbsModel,"symetric","MPLE") \%>\% Potentials
}
\author{
Victor Freguglia Souza
}
