% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gibbsMPL.R
\name{gibbsMPL}
\alias{gibbsMPL}
\title{Maximum Pseudo-Likelihood potentials estimate.}
\usage{
gibbsMPL(X, gModel, type = "symetric", ...)
}
\arguments{
\item{X}{a matrix, realization of a GibbsModel.}

\item{gModel}{A GibbsModel object, vMat and cMat value should be NULL. If the object has non-NULL vMat or V values, they'll be ignored.}

\item{type}{Wich type of Energy function to be considered. "general" indicates all potentials are free. "symetric" indicates differences of the same absolute value have the same potential value. "equal" indicates all non-zero differences potentials have the same value.}

\item{...}{arguments to be passed to the optim() function during maximization.}
}
\value{
A GibbsModel object with the estimated vMat and V potentials values.
}
\description{
Potentials estimate via Maximum Pseudo-Lilelihood.
}
