% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GibbsSamplerHMRF.R
\name{GibbsSamplerHMRF}
\alias{GibbsSamplerHMRF}
\title{Gibbs Sampler for Hidden Markov Random Field model.}
\usage{
GibbsSamplerHMRF(Y, gModel, imu, isigma2, iseg, coeftab, size = 10000)
}
\arguments{
\item{Y}{The observed values (matrix).}

\item{gModel}{a GibbsModel object.}

\item{imu}{a vector indicating initial mean values.}

\item{iseg}{initial segmentation.}

\item{isigma}{a vector indicating initial SD values.}

\item{X}{a matrix with n (number of observations) rows and p (number of fourier basis functions) columns. Use DeisgnMat() with the result of betaPLS() to obtain this matrix.}
}
\description{
Gibbs Sampler for Hidden Markov Random Field model.
}
