% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionMap.R
\name{InteractionMap}
\alias{InteractionMap}
\title{Interaction Map}
\usage{
InteractionMap(X, window_size = 10)
}
\description{
Calculates distances between observed (normalized) GLDH and IRF expected probabilities.
}
\author{
Victor Freguglia Souza
}
