% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialCorMap.R
\name{PartialCorMap}
\alias{PartialCorMap}
\title{Partial Correlation Map}
\usage{
PartialCorMap(IMG, window_size = 10, included_list = data.frame(X =
  numeric(0), Y = numeric(0)))
}
\arguments{
\item{IMG}{An image in a matrix object.}

\item{window_size}{Size of the windows to compute partial auto-correlations}

\item{included_list}{a data.frame with columns named \code{X} and \code{Y} and relative positions to condition on when computing the auto-correlations.}
}
\value{
A data.frame with absolute conditional auto-correlations for each relative position.
}
\description{
Computes conditional auto-correlations given a set of positions to condition in
}
\author{
Victor Freguglia Souza
}
