% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialCorSelection.R
\name{PartialCorSelection}
\alias{PartialCorSelection}
\title{Interaction structure selection based on partial auto-correlations}
\usage{
PartialCorSelection(IMG, window_size = 10, thr = 0.1)
}
\arguments{
\item{IMG}{An image in a matrix object.}

\item{window_size}{Size of the window for candidates to be included in the interaction structure.}

\item{thr}{A thresholding parameter. When all absolute conditional auto-correlations not included in the model are belo \code{thr}, the selection stops.}
}
\value{
A matrix with selected relative positions, ready to be used as \code{cMat} of GibbsModel.
}
\description{
Interaction structure selection based on partial auto-correlations
}
\author{
Victor Freguglia Souza
}
