% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotIMap.R
\name{plotIMap}
\alias{plotIMap}
\title{Visualization for Interaction Maps}
\usage{
plotIMap(iMap)
}
\arguments{
\item{iMap}{a DataFrame object with an Interaction Map (result of the InteractionMap() function)}
}
\description{
Plots the interaction map.
}
\author{
Victor Freguglia Souza
}
