% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstructPLS.R
\name{reconstructPLS}
\alias{reconstructPLS}
\title{Reconstruct an image based on frequencies coefficients.}
\usage{
reconstructPLS(beta_df, img_dim)
}
\arguments{
\item{beta_df}{the data.frame representation of frequency coefficients. Result of betaPLS() or buildFrequencyDf().}

\item{img_dim}{Dimensions of the image.}
}
\description{
Reconstruct an image based on frequencies coefficients.
}
\author{
Victor Freguglia Souza
}
