% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGibbsRF.R
\name{rGibbsRF}
\alias{rGibbsRF}
\title{Gibbs Random Field simulation.}
\usage{
rGibbsRF(gModel, macrosteps = 50, initial = NULL, dim = c(150, 100))
}
\arguments{
\item{gModel}{A GibbsModel object specifying wich model to sample from.}

\item{macrosteps}{The number of times each pixel will be simulated from it's conditional distribution.}

\item{initial}{A matrix indicating the initial object to start sampling from. If NULL, start with each pixel drawn from an independent discrete uniform distribution.}

\item{dim}{Used when the initial matrix is NULL. Indicates the dimension of the field to be sampled (x-axis first).}
}
\value{
A matrix, realization from the specified Gibbs Model.
}
\description{
Simulates a realization of a GibbsModel object via Gibbs Sampler.
}
