% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocAp.R
\name{StocAp}
\alias{StocAp}
\title{Estimation of potentials values via Stochastic Approximation.}
\usage{
StocAp(X, gModel, type, initial = "MPLE", MC_size = 1, iter = 10000,
  macrosteps = 40, print_sample = TRUE, a0 = 1, a1 = 1/20,
  a2 = 3)
}
\arguments{
\item{X}{The observed random field (matrix).}

\item{gModel}{A GibbsModel object (V and vMat can be NULL).}

\item{type}{Model type. Either "general", "symetric" or "equal". Check GibbsMPLE for details.}

\item{initial}{Wich method to get initial estimates. Currently only "MPLE" available.}

\item{MC_size}{Number of fields to sample at each step.}

\item{iter}{Maximum number of iterations.}

\item{macrosteps}{Number of macrosteps per random field simulation.}

\item{print_sample}{Indicates if the a sample of the current model should be printed each 10 steps.}

\item{a0, a1, a2}{constants to define size of each step. The gradient vector is multiplied by a0/(i*a1 + a2)}
}
\value{
A GibbsModel object estimated potentials.
}
\description{
Estimates potentials vias Stochastic Approximation algorithm.
}
\examples{
StocAp(example.X,example.GibbsModel,"symetric",MC_size = 5)
}
\author{
Victor Freguglia Souza
}
