% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structurePlot.R
\name{structurePlot}
\alias{structurePlot}
\title{Interaction Structure plot}
\usage{
structurePlot(cMat)
}
\arguments{
\item{cMat}{The matrix with interacting positions.}
}
\description{
Interaction Structure plot
}
\author{
Victor Freguglia Souza
}
