% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThreshIMap.R
\name{ThreshIMap}
\alias{ThreshIMap}
\title{Threshold Interaction Map}
\usage{
ThreshIMap(iMap, k)
}
\arguments{
\item{iMap}{a DataFrame with an Interaction Map (see function InteractionMap() for details).}

\item{k}{the thresholding parameter. The cut value is set as MD + k*SD. MD being the mean distance in the map, and SD their sample deviation.}
}
\value{
a DataFrame with the new Interaction Map with values thresholded.
}
\description{
Removes from Interaction Map all relative positions with lower Distance than a threshold based on k.
}
\author{
Victor Freguglia Souza
}
