% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Potentials.R
\name{Potentials}
\alias{Potentials}
\title{GibbsModel to DataFrame conversion.
Data Frame representation of GibbsModel object for better visualization.}
\usage{
Potentials(gModel)
}
\arguments{
\item{gModel}{A GibbsModel object with non-NULL potentials.}
}
\value{
A data.frame object with potentials values for each difference (column) and each relative position (row).
}
\description{
GibbsModel to DataFrame conversion.
Data Frame representation of GibbsModel object for better visualization.
}
